# backend-api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AI Receipt Tracker API",
    "description": "API for managing receipts and transactions with AI-powered data extraction using Google Gemini",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/receipts/upload": {
      "post": {
        "tags": ["Receipts"],
        "summary": "Upload receipt for AI processing (Step 1)",
        "description": "Upload a receipt image (file or URL) for AI-powered data extraction using Google Gemini. Returns extracted data for user confirmation. Does NOT save to transactions table until /receipts/confirm is called.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "Receipt image file (PNG, JPG, JPEG, GIF, WebP, HEIC)"
                  }
                }
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["image_url"],
                "properties": {
                  "image_url": {
                    "type": "string",
                    "format": "uri",
                    "example": "https://example.com/receipt.jpg",
                    "description": "Public URL of the receipt image"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Receipt uploaded and data extracted successfully. Requires user confirmation.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Receipt uploaded and data extracted successfully. Please review and confirm."
                    },
                    "receipt_id": {
                      "type": "integer",
                      "example": 42
                    },
                    "extracted_data": {
                      "$ref": "#/components/schemas/ExtractedReceiptData"
                    },
                    "requires_confirmation": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid file or missing data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "AI extraction failed or server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/receipts/confirm": {
      "post": {
        "tags": ["Receipts"],
        "summary": "Confirm receipt data (Step 2)",
        "description": "Confirm and save validated receipt data to transactions table. User can correct extracted data before saving.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["receipt_id", "vendor_name", "total_amount", "transaction_date"],
                "properties": {
                  "receipt_id": {
                    "type": "integer",
                    "example": 42,
                    "description": "ID of the receipt to confirm"
                  },
                  "vendor_name": {
                    "type": "string",
                    "example": "Starbucks Coffee"
                  },
                  "receipt_number": {
                    "type": "string",
                    "example": "INV-12345"
                  },
                  "total_amount": {
                    "type": "number",
                    "format": "float",
                    "example": 45.99
                  },
                  "currency": {
                    "type": "string",
                    "example": "USD",
                    "default": "USD"
                  },
                  "transaction_date": {
                    "type": "string",
                    "format": "date",
                    "example": "2024-12-05"
                  },
                  "payer_name": {
                    "type": "string",
                    "example": "John Doe"
                  },
                  "description": {
                    "type": "string",
                    "example": "Coffee purchase"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Receipt confirmed and transaction saved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Receipt confirmed and transaction saved successfully."
                    },
                    "receipt_id": {
                      "type": "integer",
                      "example": 42
                    },
                    "transaction_id": {
                      "type": "integer",
                      "example": 42
                    },
                    "transaction_data": {
                      "$ref": "#/components/schemas/Transaction"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid data or receipt already processed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Receipt not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/receipts/cancel": {
      "post": {
        "tags": ["Receipts"],
        "summary": "Cancel pending receipt",
        "description": "Cancel a receipt that is pending confirmation. Marks receipt as cancelled.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["receipt_id"],
                "properties": {
                  "receipt_id": {
                    "type": "integer",
                    "example": 42,
                    "description": "ID of the receipt to cancel"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Receipt cancelled successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Receipt cancelled successfully."
                    },
                    "receipt_id": {
                      "type": "integer",
                      "example": 42
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Cannot cancel processed receipt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Receipt not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/receipts/pending": {
      "get": {
        "tags": ["Receipts"],
        "summary": "Get pending receipts",
        "description": "Retrieve all receipts pending confirmation for the authenticated user",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Pending receipts retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "pending_receipts": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PendingReceipt"
                      }
                    },
                    "count": {
                      "type": "integer",
                      "example": 3
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/receipts/transactions": {
      "get": {
        "tags": ["Receipts"],
        "summary": "Get user transactions",
        "description": "Retrieve all transactions for the authenticated user",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Transactions retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "transactions": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Transaction"
                      }
                    },
                    "count": {
                      "type": "integer",
                      "example": 10
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token in the format: Bearer {token}"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Error message"
          },
          "error": {
            "type": "string",
            "example": "Detailed error information"
          }
        }
      },
      "ExtractedReceiptData": {
        "type": "object",
        "properties": {
          "vendor_name": {
            "type": "string",
            "example": "Starbucks Coffee"
          },
          "receipt_number": {
            "type": "string",
            "example": "INV-12345"
          },
          "total_amount": {
            "type": "number",
            "format": "float",
            "example": 45.99
          },
          "currency": {
            "type": "string",
            "example": "USD"
          },
          "transaction_date": {
            "type": "string",
            "format": "date",
            "example": "2024-01-15"
          },
          "payer_name": {
            "type": "string",
            "example": "John Doe"
          }
        }
      },
      "Transaction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "receipt_id": {
            "type": "integer",
            "example": 1
          },
          "vendor_name": {
            "type": "string",
            "example": "Starbucks Coffee"
          },
          "receipt_number": {
            "type": "string",
            "example": "INV-12345"
          },
          "total_amount": {
            "type": "string",
            "example": "45.99"
          },
          "currency": {
            "type": "string",
            "example": "USD"
          },
          "transaction_date": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T00:00:00"
          },
          "payer_name": {
            "type": "string",
            "example": "John Doe"
          },
          "payer_id": {
            "type": "integer",
            "example": 1
          },
          "description": {
            "type": "string",
            "example": "Transaction processed from Receipt ID 1"
          },
          "image_url": {
            "type": "string",
            "example": "uploaded://receipt.jpg"
          },
          "status": {
            "type": "string",
            "example": "PROCESSED"
          }
        }
      },
      "PendingReceipt": {
        "type": "object",
        "properties": {
          "receipt_id": {
            "type": "integer",
            "example": 42
          },
          "image_url": {
            "type": "string",
            "example": "uploaded://receipt.jpg"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2024-12-05T10:30:00"
          },
          "extracted_data": {
            "$ref": "#/components/schemas/ExtractedReceiptData"
          }
        }
      },
...
    },
...
  }
}